+++ packages/utils/src/dataknobs_utils/__init__.py
@@ -19,7 +19,7 @@ from dataknobs_utils import (
     xml_utils,
 )
 
-__version__ = "2.0.1"
+__version__ = "2.0.2"
 
 __all__ = [
     "elasticsearch_utils",