+++ packages/fsm/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.4.2 - 2026-08-26
+
 ### Fixed
 
 - **`InMemoryStorage` no longer injects connection parameters the memory